refactor: user CRUD auth 적용

+++ src/main/java/kpl/fiml/user/application/UserService.java
@@ -4,7 +4,6 @@ import kpl.fiml.global.jwt.JwtTokenProvider;
 import kpl.fiml.user.domain.User;
 import kpl.fiml.user.domain.UserRepository;
 import kpl.fiml.user.dto.*;
-import kpl.fiml.user.vo.EmailVo;
 import kpl.fiml.user.vo.PasswordVo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -41,10 +40,16 @@ public class UserService {
     }
 
     @Transactional
-    public UserUpdateResponse updateUser(Long userId, UserUpdateRequest request) {
-        User user = getById(userId);
+    public UserUpdateResponse updateUser(Long id, Long loginUserId, UserUpdateRequest request) {
+        User user = getById(id);
+        User loginUser = getById(loginUserId);
+
+        if (!user.isSameUser(loginUser)) {
+            throw new IllegalArgumentException("사용자 수정 권한이 없습니다.");
+        }
+
         String encryptPassword = request.getPassword();
-        if(!encryptPassword.isBlank()) {
+        if (!encryptPassword.isBlank()) {
             encryptPassword = validateAndEncryptPassword(request.getPassword());
         }
         user.updateUser(request.getName(), request.getBio(), request.getProfileImage(), request.getEmail(), encryptPassword, request.getContact());
@@ -52,15 +57,25 @@ public class UserService {
         return UserUpdateResponse.of(user.getId(), user.getName(), user.getBio(), user.getProfileImage(), user.getEmail(), user.getPassword(), user.getContact());
     }
 
-    public UserDto findById(Long userId) {
+    public UserDto findById(Long userId, Long loginUserId) {
         User findUser = getById(userId);
+        User loginUser = getById(loginUserId);
+
+        if (!findUser.isSameUser(loginUser)) {
+            throw new IllegalArgumentException("마이페이지 접근 권한이 없습니다.");
+        }
 
         return UserDto.of(findUser.getId(), findUser.getName(), findUser.getBio(), findUser.getProfileImage(), findUser.getEmail(), findUser.getContact(), findUser.getCash(), findUser.getCreatedAt(), findUser.getUpdatedAt());
     }
 
     @Transactional
-    public UserDeleteResponse deleteById(Long userId) {
+    public UserDeleteResponse deleteById(Long userId, Long loginUserId) {
         User user = getById(userId);
+        User loginUser = getById(loginUserId);
+
+        if (!user.isSameUser(loginUser)) {
+            throw new IllegalArgumentException("사용자 삭제 권한이 없습니다.");
+        }
         user.delete();
 
         return UserDeleteResponse.of(user.getId());
